Adds relationships
Builds out order/show view a bit

+++ database/migrations/2016_11_05_220526_create_order_lines_table.php
@@ -24,7 +24,7 @@ class CreateOrderLinesTable extends Migration
             $table->integer('xlarge_qty');
             $table->integer('xxlarge_qty');
             $table->integer('xxxlarge_qty');
-            $table->integer('line_price');
+            $table->double('line_price');
             $table->timestamps();
         });
     }
